requirements.txt
bs4
openai
gspread
google-auth 
google-auth-oauthlib